+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ Version 1.1.1 (2021-03-17)
 --------------------------
 1. The implementations of ``sc.odict()`` and ``sc.objdict()`` have been updated, to allow for more flexible use of the ``defaultdict`` argument, including better nesting and subclassing.
 2. A new ``serial`` argument has been added to ``sc.parallelize()`` to allow for quick debugging.
-3. Legacy support for ``dill`` and Python 2 has been removed from ``sc.loadobj()`` and ``sc.saveobj()``.
+3. Legacy support for Python 2 has been removed from ``sc.loadobj()`` and ``sc.saveobj()``.
 4. A fallback method for ``sc.gitinfo()`` (based on ``gitpython``) has been added, in case reading from the filesystem fails.